move auto aspect functions to own module
